requirements.txt
discord.py
asyncio
presence
GitPython
pysqlite3